+++ src/pomodouroboros/model/__init__.py
@@ -1,0 +1,45 @@
+"""
+Model v2.
+
+Spec:
+
+    - https://github.com/glyph/Pomodouroboros/issues/33
+
+What events can occur?
+
+    - time passes, which results in
+
+        - pomodoro progresses
+
+        - pomodoro ends
+
+        - break starts
+
+        - break ends
+
+    - user adds a new intention to the set of available ones
+
+    - user sets an intention from the created set
+
+        - this is different than the current system because expressing an
+          intention always results in a new pomodoro (unless one's currently
+          running of course)
+
+    - user evaluates their intentionality
+
+3 kinds of intervals
+
+    - pomodoro
+
+    - break
+
+    - grace period
+
+Okay so what is the *alignment* on these grace periods?  Do they begin at the
+time when they are originally scheduled to start, or do they start only after
+the intention is set?
+
+I think it would make the most sense to have everything aligned up front, so
+you can have some visibility into the future, but to have the grace period
+itself.
+"""
