+++ src/pomodouroboros/model/debugger.py
@@ -1,0 +1,9 @@
+def debug(*x: object) -> None:
+    """
+    Emit some messages while debugging.
+    """
+    if 0:
+        print(*x)
+
+
+
